utils: Move MakeRequest to utils
Refactor MakeRequest function to be a reusable api utility
